Fix preview and publish button being permanently disabled
The message related to this in the logs was:

Warning: Prop `disabled` did not match. Server: "" Client: "false"

Updating rum and React changed how hydration of server generated DOM works, and
having props that disagree was not working at all. Leaving the prop out on the
server seems to fix the issue.